InsertTerminalsPandas/requirements.txt
matplotlib==3.5.1
click<8.1
colorama==0.4.4
cycler==0.11.0
Cython==0.29.28
et-xmlfile==1.1.0
fonttools==4.33.3
importlib-metadata==4.11.3
install==1.3.5
itsdangerous==2.1.2
kiwisolver==1.4.2
MarkupSafe==2.1.1
numpy==1.21.6
openpyxl==3.0.9
packaging==21.3
pandas==1.3.5
Pillow==9.1.0
pyparsing==3.0.8
python-dateutil==2.8.2
pytz==2022.1
scipy==1.7.3
seaborn==0.11.2
Shapely==1.8.1.post1
six==1.16.0
Werkzeug==2.1.1
zipp==3.8.0
xlwings==0.27.7
streamlit-option-menu==0.3.2
streamlit-aggrid == 0.3.2
networkx==2.8
XlsxWriter==3.0.3

